+++ src/README.md
@@ -38,6 +38,7 @@ highlight.c	| syntax highlighting
 indent.c	| C and Lisp indentation
 insexpand.c	| Insert mode completion
 mark.c		| marks
+map.c		| mapping and abbreviations
 mbyte.c		| multi-byte character handling
 memfile.c	| storing lines for buffers in a swapfile
 memline.c	| storing lines for buffers in memory
