+++ src/proto.h
@@ -98,6 +98,7 @@ extern int _stricoll(char *a, char *b);
 # include "list.pro"
 # include "blob.pro"
 # include "main.pro"
+# include "map.pro"
 # include "mark.pro"
 # include "memfile.pro"
 # include "memline.pro"
